requirements.txt
certifi==2023.5.7
charset-normalizer==2.0.12
idna==3.4
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.16
